Remove a bunch of language codes from the unisubs mapping
We want to have these codes in the system for other standards to use,
but we don't want them to be available for "normal" language codes
(e.g. setting the primary audio language or adding a new subtitle
language).

This change is the start of a new policy for the unisubs mapping.
Before it used to contain all the internal codes, but now it's a
subset.

Added LanguageCode.to_internal() to get the internal language code.  I
think we should use that where we used to use encode('unisubs').  Also
allow creating LanguageCodes from internal language codes (by using
standard="internal").  Lastly, allow "internal" for
get_language_name_mapping().

I think this API could use a bit of cleanup.  I made a ticket for that
because there's not really time to do it now.